feat(测试跟踪): 测试计划关联场景支持全选
--story=1008184 --user=陈建星 测试计划优化 https://www.tapd.cn/55049933/s/1198664
