+++ backend/src/main/java/io/metersphere/base/mapper/ext/ExtApiScenarioMapper.java
@@ -93,5 +93,5 @@ public interface ExtApiScenarioMapper {
 
     List<ApiScenarioWithBLOBs> selectRepeatByBLOBs(@Param("names") List<String> names, @Param("projectId") String projectId);
 
-
+    List<String> selectRelevanceIdsByQuery(@Param("request") BaseQueryRequest query);
 }
