+++ backend/src/main/java/io/metersphere/track/service/TestPlanScenarioCaseService.java
@@ -5,6 +5,7 @@ import com.alibaba.fastjson.JSONObject;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import io.metersphere.api.dto.EnvironmentType;
+import io.metersphere.api.dto.ScenarioEnv;
 import io.metersphere.api.dto.automation.*;
 import io.metersphere.api.service.ApiAutomationService;
 import io.metersphere.api.service.ApiScenarioReportService;
@@ -19,11 +20,7 @@ import io.metersphere.commons.constants.ExecuteResult;
 import io.metersphere.commons.constants.ProjectApplicationType;
 import io.metersphere.commons.constants.TestPlanTestCaseStatus;
 import io.metersphere.commons.exception.MSException;
-import io.metersphere.commons.utils.PageUtils;
-import io.metersphere.commons.utils.Pager;
-import io.metersphere.commons.utils.ServiceUtils;
-import io.metersphere.commons.utils.TestPlanUtils;
-import io.metersphere.controller.request.OrderRequest;
+import io.metersphere.commons.utils.*;
 import io.metersphere.controller.request.ResetOrderRequest;
 import io.metersphere.dto.MsExecResponseDTO;
 import io.metersphere.dto.ProjectConfig;
@@ -32,9 +29,9 @@ import io.metersphere.log.vo.OperatingLogDetails;
 import io.metersphere.service.ProjectApplicationService;
 import io.metersphere.service.ProjectService;
 import io.metersphere.track.dto.*;
+import io.metersphere.track.request.testcase.ApiCaseRelevanceRequest;
 import io.metersphere.track.request.testcase.TestPlanScenarioCaseBatchRequest;
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
@@ -140,6 +137,57 @@ public class TestPlanScenarioCaseService {
         }
         Page<Object> page = PageHelper.startPage(goPage, pageSize, true);
         return PageUtils.setPageInfo(page, apiAutomationService.list(request));
+    }
+
+    public void relevance(ApiCaseRelevanceRequest request) {
+        if (testPlanService.isAllowedRepeatCase(request.getPlanId())) {
+            request.getCondition().setNotInTestPlan(false);
+        } else {
+            request.getCondition().setNotInTestPlan(true);
+        }
+
+        apiAutomationService.buildApiCaseRelevanceRequest(request);
+
+        Map<String, List<String>> mapping = request.getMapping();
+        Map<String, String> envMap = request.getEnvMap();
+        Set<String> set = mapping.keySet();
+        List<String> relevanceIds = request.getIds();
+        Collections.reverse(relevanceIds);
+        String envType = request.getEnvironmentType();
+        String envGroupId = request.getEnvGroupId();
+        if (set.isEmpty()) {
+            return;
+        }
+        Long nextOrder = ServiceUtils.getNextOrder(request.getPlanId(), extTestPlanScenarioCaseMapper::getLastOrder);
+        for (String id : relevanceIds) {
+            Map<String, String> newEnvMap = new HashMap<>(16);
+            List<String> list = mapping.get(id);
+            if (CollectionUtils.isEmpty(list)) {
+                ScenarioEnv scenarioEnv = apiAutomationService.getApiScenarioProjectId(id);
+                list = new ArrayList<>(scenarioEnv.getProjectIds());
+            }
+            list.forEach(l -> newEnvMap.put(l, envMap == null ? "" : envMap.getOrDefault(l, "")));
+            TestPlanApiScenario testPlanApiScenario = new TestPlanApiScenario();
+            testPlanApiScenario.setId(UUID.randomUUID().toString());
+            testPlanApiScenario.setCreateUser(SessionUtils.getUserId());
+            testPlanApiScenario.setApiScenarioId(id);
+            testPlanApiScenario.setTestPlanId(request.getPlanId());
+            testPlanApiScenario.setCreateTime(System.currentTimeMillis());
+            testPlanApiScenario.setUpdateTime(System.currentTimeMillis());
+            String environmentJson = JSON.toJSONString(newEnvMap);
+            if (StringUtils.equals(envType, EnvironmentType.JSON.name())) {
+                testPlanApiScenario.setEnvironment(environmentJson);
+                testPlanApiScenario.setEnvironmentType(EnvironmentType.JSON.name());
+            } else if (StringUtils.equals(envType, EnvironmentType.GROUP.name())) {
+                testPlanApiScenario.setEnvironmentType(EnvironmentType.GROUP.name());
+                testPlanApiScenario.setEnvironmentGroupId(envGroupId);
+                // JSON类型环境中也保存最新值
+                testPlanApiScenario.setEnvironment(environmentJson);
+            }
+            testPlanApiScenario.setOrder(nextOrder);
+            nextOrder += ServiceUtils.ORDER_STEP;
+            testPlanApiScenarioMapper.insert(testPlanApiScenario);
+        }
     }
 
     public int delete(String id) {
